Split methods to different classes
